refactor: extract value and string helpers
